Turned fs::CountFilesRecursive() into fs::CountFilesAndDirsRecursive() Fixed Contains in Properties screen Fixed warning in Browser::GetFilesSize() Fixed Browser::CountMultipleFilesType() Fixed float instead of u32 number in start number of files in copy_layout.cpp

+++ source/layouts/copy_layout.cpp
@@ -41,7 +41,7 @@ namespace ui
         else
             this->item = " item...";
 
-        this->copyTextHeader->SetText(this->action + std::to_string(this->copyProgressBar->GetProgress()) + " of " + std::to_string(this->numberOfElements) + this->item);
+        this->copyTextHeader->SetText(this->action + std::to_string((u32) this->copyProgressBar->GetProgress()) + " of " + std::to_string(this->numberOfElements) + this->item);
         this->copyProgressBar->SetMaxValue(this->numberOfElements);
         this->copyProgressBar->ClearProgress();
         app->CallForRender();
